rename mock folder
